fix: route tools_*.py imports through server._misc for layer compliance
REQ-IMP-003 / REQ-ARCH-003 restrict tools_*.py to importing from
{core, pipeline, server, config, fleet}. The initial refactor routed
passthrough re-exports directly to autoskillit.execution and
autoskillit.workspace, violating these arch constraints.

Fixes:
- Add re-exports to _misc.py for execution/workspace symbols
- Move _build_hook_diagnostic_warning from tools_kitchen.py to _misc.py
- Redirect all tools_*.py imports through server._misc
- Patch targets in test_tools_git.py → server._misc.resolve_remote_name
- Promote _refresh_quota_cache to module-level import in _misc.py
- Update schema version allowlist line numbers
- Document _notify.py and _misc.py in CLAUDE.md
- Fix stale server/helpers comment in _type_constants.py
